finalizando o autoinsert

+++ CadAuto/autoinsert.py
@@ -3,46 +3,56 @@ import shutil
 import os
 import pyperclip as pc
 import csv
-from time import sleep 
+from time import sleep
 
 
 class autoinsert:
 
     def Auto_Insert(caminho):
-        
+
         filename = os.path.basename(caminho)
         print(filename)
-       
+
         destino_relativo = "Sistema-de-armazenamento-de-produtos/CadAuto"
         destino_absoluto = os.path.abspath(destino_relativo)
 
-        shutil.copy(caminho, destino_absoluto)
-        print(f"Arquivo copiado em {destino_absoluto}")
+        if not os.path.exists(destino_absoluto):
+            shutil.copy(caminho, destino_absoluto)
+            print(f"Arquivo copiado em {destino_absoluto}")
+        else:
+            print(f"Arquivo já existe em {destino_absoluto}")
 
-        with open(caminho,mode='r',encoding='latin1') as arquivo:
+        with open(caminho, mode="r", encoding="latin1") as arquivo:
             leitor = csv.reader(arquivo)
 
             next(leitor, None)
 
             for row in leitor:
-                ag.click(249,148,duration=1)
+                sleep(0.5)
                 nome = row[0]
                 pc.copy(nome)
-                ag.hotkey('ctrl','v')
-                ag.click(330,174,duration=1)
+                ag.hotkey("ctrl", "v")
+                ag.press("tab")
+                sleep(0.5)
                 quantidade = int(row[1])
                 pc.copy(quantidade)
-                ag.hotkey('ctrl','v')
-                ag.click(322,200,duration=1)
+                ag.hotkey("ctrl", "v")
+                ag.press("tab")
+                sleep(0.5)
                 preco = float(row[2])
                 pc.copy(preco)
-                sleep(1)
-                ag.hotkey('ctrl','v')
-                ag.press('enter')
+                ag.hotkey("ctrl", "v")
+                sleep(0.5)
+                ag.press("enter")
+                sleep(0.5)
+                ag.press("enter")
+                ag.press("tab")
+                ag.press("tab")
+                ag.press("tab")
+                ag.press("tab")
+                ag.press("tab")
+                ag.press("tab")
+                ag.press("tab")
+                ag.press("tab")
 
-                
-
-
-    
-        
         pass
